ifconfig: Fix flags value
ifr.ifr_flags's type is short. When short type convert to u32,its expand bit will 1 if short value is negative.

+++ tools/ifconfig/network_interface.cc
@@ -147,7 +147,7 @@ interface::interface(const std::string& iface_name)
     bsd_ifreq ifr;
 
     if (do_ifr(sock, name, SIOCGIFFLAGS, &ifr) == 0) {
-        uflags = ((uint32_t) ifr.ifr_flagshigh << 16) | ifr.ifr_flags;
+        uflags = ((uint32_t) ifr.ifr_flagshigh << 16) | (uint16_t) ifr.ifr_flags;
         flags = flags2str(uflags);
     }
